pyproject.toml
[project]
name = "video-transcription"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper",
    "boto3",
    "python-dotenv",
]
